Move tests folder inside, add test api
